# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KangaruRide API
  version: 1.0.0
  description: >
    Enterprise transport management platform API (v1). All responses use the
    envelope defined in AGENTS.md API Standards: successes carry
    `{success: true, message, data}` (plus `meta` where documented), and
    failures carry `{success: false, code, message, errors}` with a stable
    machine-readable `code`.
paths:
...
  /api/v1/me/offers:
    get:
      summary: Jobs currently offered to the signed-in driver.
      description: >
        Live offers only. An offer whose clock ran out while the handset was
        in a dead zone is absent rather than listed — rendering it would give
        the driver a countdown reading a negative number and a button whose
        only outcome is a 409.


        403 NOT_A_DRIVER when the account is not linked to a driver profile.
      responses:
        '200':
          description: Offers in front of this driver, soonest to expire first.
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                additionalProperties: false
                properties:
                  success: { enum: [true] }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/DispatchOffer' }
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  responses:
    Error:
      description: >
        Any failure. `code` is stable and machine-readable — clients branch
        on it, never on message text (AGENTS.md API Standards).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    EmptyableStringListMap:
      description: >
        Validation errors — a map of field name to messages. PHP serialises
        the empty map as [], hence the anyOf.
      anyOf:
        - type: array
          maxItems: 0
        - type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    ErrorCode:
      description: Every case of App\Enums\ErrorCode. Never removed, never repurposed.
      enum:
        - VALIDATION_FAILED
        - INVALID_CREDENTIALS
        - MFA_REQUIRED
        - MFA_ENROLMENT_REQUIRED
        - MFA_CHALLENGE_INVALID
        - MFA_CODE_INVALID
        - MFA_ALREADY_ENROLLED
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - SERVER_ERROR
        - INVALID_TRIP_TRANSITION
        - INVALID_BOOKING_TRANSITION
        - VEHICLE_UNAVAILABLE
        - DRIVER_UNAVAILABLE
        - REPORT_TOO_LARGE
        - EXPORT_NOT_READY
        - EXPORT_EXPIRED
        - ALLOCATION_CONFLICT
        - VEHICLE_EXCLUSIVELY_ALLOCATED
        - INVALID_ORDER_REQUEST_TRANSITION
        - RATE_CARD_NOT_CONFIGURED
        - TRIP_NOT_INVOICEABLE
        - TRIP_NOT_INVOICEABLE_WALK_IN
        - TRIP_ALREADY_INVOICED
        - IDEMPOTENCY_KEY_REUSED
        - CREDIT_NOTE_EXCEEDS_INVOICE
        - ROLE_IN_USE
        - ORDERING_PAUSED
        - MAIL_DELIVERY_FAILED
        - DRIVER_ACCOUNT_CONFLICT
        - AVAILABILITY_ALREADY_ANSWERED
        - AUTOMATIC_DISPATCH_DISABLED
        - NO_DISPATCH_CANDIDATE
        - NOT_A_DRIVER
        - NOT_ON_DUTY
        - OFFER_NO_LONGER_OPEN
        - OFFER_HAS_NO_VEHICLE
        - TOKEN_SCOPE_EXCEEDED
...
    ErrorEnvelope:
      type: object
      required: [success, code, message, errors]
      additionalProperties: false
      properties:
        success: { enum: [false] }
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
        errors: { $ref: '#/components/schemas/EmptyableStringListMap' }

    # -------------------------- resources ----------------------------
...
    DispatchOffer:
      description: >
        A job put in front of one driver, with a clock on it (ADR-0024 §3).


        Deliberately carries **no passenger name or phone number**. Those are
        released only after the driver accepts (ADR-0024 §7): a number handed
        to a driver who then declines is given away for nothing, and this
        payload is what a push notification is built from — which puts it on
        a lock screen readable by whoever is holding the phone.
      type: object
      required:
        - id
        - status
        - expires_at
        - expires_in_seconds
        - pickup
        - dropoff
        - service_type
        - reference
        - pickup_distance_km
        - reasons
        - vehicle_id
      additionalProperties: false
      properties:
        id: { type: integer }
        status: { enum: [offered, accepted, declined, expired, superseded] }
        expires_at: { type: string, format: date-time }
        expires_in_seconds:
          description: >
            Served alongside the timestamp so the first render is right even
            on a handset whose clock is minutes out — common enough on cheap
            Android hardware to matter, and it would otherwise show a driver
            40 seconds left on a 15-second offer.
          type: integer
          minimum: 0
        pickup: { $ref: '#/components/schemas/OfferPlace' }
        dropoff: { $ref: '#/components/schemas/OfferPlace' }
        service_type: { type: ['string', 'null'] }
        reference: { type: ['string', 'null'] }
        pickup_distance_km:
          description: Straight-line, not road distance — a ranking, not a promised ETA (ADR-0020 §3).
          type: ['number', 'null']
        reasons:
          description: One sentence per scoring component, so a ranking stays auditable (ADR-0020 §4).
          type: array
          items: { type: string }
        vehicle_id: { type: ['integer', 'null'] }
        vehicle_registration: { type: ['string', 'null'] }
    OfferPlace:
      type: object
      required: [label, latitude, longitude]
      additionalProperties: false
      properties:
        label: { type: ['string', 'null'] }
        latitude: { type: ['number', 'null'] }
        longitude: { type: ['number', 'null'] }
...